# swagger.yaml
openapi: 3.0.3
info:
  title: IoT Control Panel API
  version: 1.0.0
  description: >
    API cho hệ thống IoT (thiết bị + cảm biến).  
    Bao gồm quản lý thiết bị, dữ liệu cảm biến, lịch sử điều khiển, MQTT và WebSocket.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/history/control:
    get:
      summary: Lấy lịch sử điều khiển thiết bị
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
          example: 1
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
          example: 5
        - in: query
          name: deviceId
          schema: { type: string }
          example: "esp8266-001"
        - in: query
          name: action
          schema: { type: string, enum: [ON, OFF, TOGGLE] }
          example: "ON"
        - in: query
          name: startDate
          schema: { type: string, format: date-time }
          example: "2025-09-28T00:00:00Z"
        - in: query
          name: endDate
          schema: { type: string, format: date-time }
          example: "2025-09-29T23:59:59Z"
      responses:
        '200':
          description: Lịch sử điều khiển
          content:
            application/json:
              example:
                success: true
                data:
                  items:
                    - deviceId: "esp8266-001"
                      action: "ON"
                      source: "web"
                      timestamp: "2025-09-29T12:40:00Z"
                    - deviceId: "esp8266-001"
                      action: "OFF"
                      source: "mqtt"
                      timestamp: "2025-09-29T13:00:00Z"
                  pagination:
                    page: 1
                    limit: 5
                    totalPages: 3
...
  /api/control/device/{deviceId}/history:
    get:
      summary: Lấy lịch sử điều khiển của 1 thiết bị
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string }
          example: "esp8266-001"
      responses:
        '200':
          description: Lịch sử điều khiển của thiết bị
          content:
            application/json:
              example:
                success: true
                data:
                  - action: "ON"
                    timestamp: "2025-09-29T12:40:00Z"
                  - action: "OFF"
                    timestamp: "2025-09-29T13:00:00Z"